revert unnecessary extra product_list file
